\name{rvTDT.example}
\alias{rvTDT.example}
\title{Example data for rvTDT}
\description{
Example data for rvTDT.
}
\format{
  rvTDT.example contains the following objects:
  \describe{
    \item{ped}{a numeric genotype matrix of 447 individuals and 19 SNPs. 
	Each row represents a different individual, and each column represents a different SNP marker. The first 149 row are genotype of children and 150-298 rows are genotype of corresponding mather and the last 149 rows are corresponding father  }
    \item{evs}{a numeric matrix of evs information for 19 variants: rowname: varaint ID: chr_pos_ref_alt; col1: number of samples have genotype 2;col2: number of samples have genotype 1;col3: number of samples have genotype 0;col4: mean coverage at this locus }
  }
}

